climdata/requirements.txt
beautifulsoup4==4.13.5
earthengine_api==1.5.19
geemap==0.35.3
geopandas==1.0.1
google_api_python_client==2.172.0
hydra-core==1.3.2
ipdb==0.13.13
numpy==2.3.2
omegaconf==2.3.0
pandas==2.3.2
Pint==0.24.4
Pint_Pandas==0.7.1
protobuf==6.32.0
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.5
scipy==1.16.1
tqdm==4.67.1
wetterdienst==0.111.0
xarray==2025.4.0
